refactor: 提取 source helper
