+++ src/helper/hook.ts
@@ -1,9 +1,9 @@
-import path from "node:path";
 import { parse } from "@babel/parser";
-import * as t from "@babel/types";
 import traverse, { NodePath } from "@babel/traverse";
-import { getNodeJSRelativePath } from "../utils/path";
+import * as t from "@babel/types";
+import path from "node:path";
 import { FileProcesser, Sentence, SentenceType } from "../type";
+import { SourceHelper, SourceParam } from "./source";
 
 function getFunctionName(path: NodePath<t.Function>) {
   if (t.isFunctionDeclaration(path.node)) {
@@ -162,26 +162,7 @@ function isFunctionIsReactComponent(
  * 根据hook解析依赖 与 scope
  */
 export function HookHelper(
-  source: {
-    /**
-     * relative path start with './'
-     */
-    importSource: string;
-    /**
-     * 引用名称
-     */
-    name: string;
-    /**
-     * 是否为默认引用
-     * @default false
-     */
-    isDefault?: boolean;
-    /**
-     * 是否为 * as namespace
-     * @default false
-     */
-    isNamespace?: boolean;
-  },
+  source: SourceParam,
   call: {
     /**
      * 调用名
@@ -192,117 +173,18 @@ export function HookHelper(
     params?: string;
   }
 ) {
-  const isRelativePath = source.importSource.startsWith("./");
+  const sourceHelper = SourceHelper(source);
 
   function parseFile(filePath: string, fileContent: string) {
-    const moduleImportPath = isRelativePath
-      ? getNodeJSRelativePath(filePath, source.importSource)
-      : source.importSource;
+    const sourceResult = sourceHelper.parse(filePath, fileContent);
+
+    const { localImportName } = sourceResult;
 
     const ast = parse(fileContent, {
       sourceFilename: filePath,
       sourceType: "module",
       plugins: ["typescript", "jsx"],
     });
-
-    let hasImport = false;
-    let hasSpecifier = false;
-
-    let importInsertIndex = 0;
-    let importInsert = "";
-    let specifierInsertIndex: null | number = null;
-    let specifierInsert = "";
-
-    /**
-     * 本地引用名称
-     */
-    let localImportName: string | null = null;
-
-    ast.program.body.forEach((node) => {
-      if (t.isImportDeclaration(node)) {
-        importInsertIndex = node.end as number;
-
-        if (node.source.value === moduleImportPath) {
-          hasImport = true;
-
-          const firstSpecifier = node.specifiers[0];
-
-          const isDefault = t.isImportDefaultSpecifier(firstSpecifier);
-          const isNamespace = t.isImportNamespaceSpecifier(firstSpecifier);
-
-          if (source.isDefault) {
-            node.specifiers.some((specifier) => {
-              if (t.isImportDefaultSpecifier(specifier)) {
-                localImportName = specifier.local.name;
-              }
-            });
-
-            if (!localImportName) {
-              if (isNamespace) {
-                console.warn(
-                  `namespace specifier cannot be used with other specifier at the same time, found in ${filePath}`
-                );
-                localImportName = `${firstSpecifier.local.name}.default`;
-              } else {
-                specifierInsertIndex = node.specifiers[0].start as number;
-                specifierInsert = `${source.name}, `;
-              }
-            }
-          } else if (source.isNamespace) {
-            node.specifiers.some((specifier) => {
-              if (t.isImportNamespaceSpecifier(specifier)) {
-                localImportName = specifier.local.name;
-              }
-            });
-
-            if (!localImportName) {
-              console.warn(
-                `namespace specifier cannot be used with other specifier at the same time, found in ${filePath}`
-              );
-
-              /**
-               * 将 hasImport 设置为 false，重复插入 import * as xx;
-               */
-              hasImport = false;
-            }
-          } else {
-            node.specifiers.forEach((specifier) => {
-              if (t.isImportSpecifier(specifier)) {
-                specifierInsertIndex = specifier.end as number;
-                const importedName = t.isIdentifier(specifier.imported)
-                  ? specifier.imported.name
-                  : specifier.imported.value;
-                if (importedName === source.name) {
-                  localImportName = specifier.local.name;
-                }
-              }
-            });
-
-            if (!localImportName && specifierInsertIndex === null) {
-              if (isNamespace) {
-                console.warn(
-                  `namespace specifier cannot be used with other specifier at the same time, found in ${filePath}`
-                );
-                hasImport = false;
-              } else if (isDefault) {
-                specifierInsertIndex = firstSpecifier.end;
-                specifierInsert = `, { ${source.name} }`;
-              }
-            }
-          }
-        }
-      }
-    });
-
-    if (!hasImport) {
-      importInsert = `import ${
-        source.isDefault
-          ? source.name
-          : source.isNamespace
-          ? `* as ${source.name}`
-          : `{ ${source.name} }`
-      } from "${moduleImportPath}"`;
-    }
 
     const scopes: {
       start: number;
@@ -333,7 +215,6 @@ export function HookHelper(
           },
         });
         if (desc) {
-          console.log(desc);
           switch (true) {
             case desc.hasCallUse:
             case desc.isFunctionNameStartWithUse:
@@ -356,22 +237,8 @@ export function HookHelper(
 
     return {
       matched: scopes.length !== 0,
-
-      hasImport,
-      /**
-       * import语句插入位置
-       */
-      importInsertIndex,
-      /**
-       * import语句
-       */
-      importInsert,
-      hasSpecifier,
-      /**
-       * import语句中的引用插入位置
-       */
-      specifierInsertIndex,
-      specifierInsert,
+      sourceHelper,
+      source: sourceResult,
       scopes,
       defaultHookCall,
     };
@@ -433,22 +300,9 @@ export function HookHelper(
       context.replace(replacement);
     },
     afterSentenceReplace(context: Context, sentence: Sentence) {
-      const {
-        scopes,
-        defaultHookCall,
-        hasImport,
-        importInsert,
-        importInsertIndex,
-      } = context.result;
+      const { scopes, defaultHookCall, sourceHelper, source } = context.result;
 
-      if (!hasImport) {
-        context.insert(
-          importInsertIndex,
-          importInsertIndex,
-          `\n${importInsert}`
-        );
-        context.result.hasImport = true;
-      }
+      sourceHelper.addImport(context, source);
 
       const macthedScope = scopes.find((scope) => {
         return sentence.start >= scope.start && sentence.end <= scope.end;
